backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.30.6
pydantic-settings==2.6.1
python-dotenv==1.0.1

httpx==0.28.1
